% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmdFunctions_general.R
\name{f.proast}
\alias{f.proast}
\title{Main function to perform proast analysis}
\usage{
f.proast(odt = list(), dt.list = 0, resize = FALSE, shinyInput,
  track = FALSE)
}
\arguments{
\item{odt}{list as returned by f.scan()}

\item{dt.list}{list with results from previous proast run}

\item{resize}{boolean, TRUE if the graph window should be resized;
default value is FALSE}

\item{shinyInput}{list with values assigned in the shiny app}

\item{track}{logical, if TRUE (FALSE by default) print the name of the function which is currently being run}
}
\value{
ans.all, list with all results that were obtained during the analysis
}
\description{
Main function to perform proast analysis
}

