% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performAnalysis.R
\name{fitModels}
\alias{fitModels}
\title{Fit all models for the continuous or quantal response}
\usage{
fitModels(data, shinyInput, continuousResponse, track = FALSE)
}
\arguments{
\item{data}{list, as returned from the f.scan() function;}

\item{shinyInput}{list, all parameter values as defined in the shiny UI}

\item{continuousResponse}{boolean, indicating whether the response is a
continuous covariate or not}

\item{track}{logical, if TRUE (FALSE by default) print the name of the function which is currently being run}
}
\value{
list with for each model a list with all results obtained during 
the analysis with f.proast(); attribute "modelNames" contains the full names
of all models that were fitted; if error occurs return error message
}
\description{
Fit all models for the continuous or quantal response
}

