% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performAnalysis.R
\name{summaryModels}
\alias{summaryModels}
\title{Summarize the results of all fitted models for continuous or quantal response}
\usage{
summaryModels(savedResults)
}
\arguments{
\item{savedResults}{list as returned by the function fitModels()}
}
\value{
data frame with for each model the model = model names, npar = number 
of parameters in the model, loglik = the estimated log likelihood, aic = the
estimated aic, bmd, bmdl and bmdu, respectively the estimate, lower and upper 
bound of the benchmark dose, converged = whether the fitting procedure 
ended with convergion or not; attributes minimum and maximum of all bmdl resp
bmdu values, the index and name of the best (min aic) model, the name of the 
response
}
\description{
Summarize the results of all fitted models for continuous or quantal response
}

